feat: Add Avatar component
This commit adds a new component called Avatar. The Avatar component is responsible for rendering an image with fallback text in case the image fails to load. It uses the @rn-primitives/avatar package for the avatar styling. The component accepts props such as imageURI, alt, fallback, fallbackColor, and style. It also uses the useTailwind hook for styling.